Add files via upload
What is a blockchain? essentially, a blockchain is a public database that irreversibly documents and authenticates the possession and transmission of digital assets. 
Digital currencies, like Bitcoin and Ethereum, are based on this concept. 
Blockchain is an exciting technology that you can use to transform the capabilities of your applications.